python-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
psycopg2-binary==2.9.9
pillow==10.1.0
httpx==0.25.2
aiohttp==3.9.1
numpy==1.26.2
opencv-python==4.8.1.78
groq==0.4.1
requests==2.31.0
cors==1.0.1
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2==2.9.9
